bets

+++ easybet/database/migrations/2014_10_12_000000_create_users_table.php
@@ -24,7 +24,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('wallet')->default(0);
+            $table->float('wallet')->default(0);
             $table->rememberToken();
             $table->timestamps();
             $table->boolean('admin')->nullable();
